Implement profiles for the Genesis controller
The Genesis controller can now be in two different modes if using
an 8bitdo controller.

1. MiSTer mode
- The home button sends "DOWN + MODE"
(This is because no equivalent of the home button exists
on the other controller ports and MiSTer require they button
mapping to be the same on all three ports)

- Buttons are swapped A with B and X with Y. This is such that the
position of the buttons is consistent between SNES and Genesis.

2. Normal mode

The mode can be toggled with "HOME + Z". The current mode is saved
in EEPROM and will therefore persist over power offs.